Ref to converting the NYC LION GeoDB into ESRI Shapefile
Note that the New York City's samples of Traffic Volume Count per
street-segment, that we use in this project:

  https://data.cityofnewyork.us/NYC-BigApps/Traffic-Volume-Counts-2012-2013-/p424-amsu

refer to segments in the NYC Dept. of City Planning LION
Single-Line Street database, so, since we use the Traffic Volume
Count per street-segment, we need to use the conversion of
the NYC LION GeoDB into ESRI Shapefile.