following the Scala coding style guides and conventions, at http://docs.scala-lang.org/style/
This is the second batch on other classes of the previous
commit, on the first set of classes/objects.

The purpose of this project is more attaining Data Mining
on the New York City's Traffic, but coding style guides are better
used earlier than later, and now all current work was just been
modularized in packages, classes and objects.